x86: Prepare to introduce amd_iommu code
Add few modifications to Jailhouse codebase required for amd_iommu code
to work properly. First, we allocate IOMMU page tables statically in
struct cell. Second, account for page sizes larger than 2GB in struct paging.

+++ hypervisor/include/jailhouse/paging.h
@@ -71,7 +71,7 @@ typedef pt_entry_t page_table_t;
 struct paging {
 	/** Page size of terminal entries in this level or 0 if none are
 	 * supported. */
-	unsigned int page_size;
+	unsigned long page_size;
 
 	/**
 	 * Get entry in given table corresponding to virt address.
